add auth controller and service

+++ Controllers/StorageController.cs
@@ -7,14 +7,21 @@ namespace frontend.Controllers
   public class StorageController : Controller
   {
     private readonly IStorageService RegistrationService;
-    public StorageController(IStorageService registrationService)
+    private readonly IAuthService authService;
+
+    public StorageController(IStorageService registrationService, IAuthService authService)
     {
       RegistrationService = registrationService;
+      this.authService = authService;
     }
 
     [HttpPost]
     public IActionResult Register([FromBody] Person person)
     {
+      if (!authService.IsLoggedIn())
+      {
+        return Unauthorized(new { message = "Unauthorized" });
+      }
       RegistrationService.Register(person);
       return Ok();
     }
@@ -22,6 +29,10 @@ namespace frontend.Controllers
     [HttpGet("all")]
     public IActionResult GetAll()
     {
+      if (!authService.IsLoggedIn())
+      {
+        return Unauthorized(new { message = "Unauthorized" });
+      }
       var people = RegistrationService.GetAll();
       return Ok(people);
     }
@@ -29,6 +40,10 @@ namespace frontend.Controllers
     [HttpDelete]
     public IActionResult Delete([FromQuery] Guid id)
     {
+      if (!authService.IsLoggedIn())
+      {
+        return Unauthorized(new { message = "Unauthorized" });
+      }
       RegistrationService.Delete(id);
       return Ok();
     }
